BUG FIX: last commit needed to use size_t for unpropagated_index_table_MBs, because we multiplied it by 1024*1024 and it would overflow otherwise

+++ hoffman.cc
@@ -9276,7 +9276,7 @@ struct UnpropagatedIndexTable {
 };
 
 UnpropagatedIndexTable * unpropagated_index_table = NULL;
-int unpropagated_index_table_MBs = 0;
+size_t unpropagated_index_table_MBs = 0;
 
 /* finalize_update()
  *
